Error en el recuento de líneas y reporte de errores

+++ biblioteca_dinamica.h
@@ -8,7 +8,7 @@ typedef void* BibilotecaDinamica;
 /// Por ahora tenemos la limitación de que tienen
 /// que tener un prototipo fijo de 2 int de entrada
 /// y 1 de salida.
-typedef int (*FuncionForanea)(int, int);
+typedef int (*FuncionForanea)(double , double);
 
 BibilotecaDinamica cargar_biblioteca_dinamica(char* fichero);
 void cerrar_biblioteca_dinamica(BibilotecaDinamica *bib);
